Add rows to the calendar
Rows will contain the data for the schedule as well as hold the date for
the day in the month. Tentatively using a table, not sure if this will
be cumbersome when adding items from the schedule

Calendar template currently labels all 42 cells in the month. This will be
changed later as we take into account what day the month starts on and
populating the cells that are needed.