Add markdown selection replacement in journal
This feature allows for users to apply the markdown to the text that
they have selected

Previously, clicking one of the markdown buttons would have replaced
deleted selected text and replaced it with the appropriate
markdown symbols that is parsed by React-Markdown.
This was counterintuitive for users as they had wanted
the markdown to be applied to the text that they had selected.