Change homepage date after editing journal date
Changing the journal entry date and clicking the back button
would bring users back to the old entry date. This was
unnatural as users would not see the entry that they had just edited.
After clicking the back button, the homepage now renders the edited
entry in the changed calendar date, as well as the entries that
are connected to their date

e.g User changes post from 20 May 2021 to 10 May 2021. Going to the
homepage shows all the posts from 10 May 2021 instead of 20 May 2021